rename Mint directory to Market
