+++ frontend/components/Header.tsx
@@ -22,8 +22,8 @@ export function Header() {
       <div className="flex gap-2 items-center flex-wrap">
         {IS_DEV && (
           <>
-            <Link className={buttonVariants({ variant: "link" })} to="/mint">
-              Mint Page
+            <Link className={buttonVariants({ variant: "link" })} to="/market">
+              Market Page
             </Link>
             <Link className={buttonVariants({ variant: "link" })} to="/my-collections">
               My Collections
