+++ frontend/components/NavBar.tsx
@@ -18,7 +18,7 @@ interface NavigationItem {
 // Define navigation array with proper typing
 const navigation: NavigationItem[] = [
   { name: 'Home', href: '/', current: false },
-  { name: 'Market', href: '/mint', current: true },
+  { name: 'Market', href: '/market', current: true },
 ];
 
 interface NavBarProps {
