+++ frontend/pages/Market/components/BannerSection.tsx
