+++ frontend/pages/Market/components/FAQSection.tsx
