+++ frontend/pages/Market/components/HowToMintSection.tsx
