+++ frontend/pages/Market/components/OurTeamSection.tsx
