+++ frontend/pages/Market/components/StatsSection.tsx
